Bugs are fixed - now just need to implement command lists then I can start testing.

+++ phxengine/include/phx/core/ThreadPool.h
@@ -35,8 +35,8 @@ namespace phx
 
 		void Signal(Barrier& barrier);
 
-		size_t GetThreadCount();
-		size_t GetNumCores();
+		uint32_t GetThreadCount();
+		uint32_t GetNumCores();
 	}
 
 }
